% Generated by roxygen2 (4.0.2): do not edit by hand
\name{choose_maxdist_option}
\alias{choose_maxdist_option}
\title{Function for translating maxdiff option into value}
\usage{
choose_maxdist_option(distM, maxDist = "Inf")
}
\arguments{
\item{distM}{A distance matrixe calculated by one of the functions from
  \code{distanceFunctions}.}

\item{maxDist}{The maxDist option to decide upon.}
}
\description{
Function for translating maxdiff option into value
}

