% Generated by roxygen2 (4.0.2): do not edit by hand
\name{diffr}
\alias{diffr}
\alias{textAlign}
\title{A function for aligning text and measuring its distance.}
\usage{
diffr(text1, text2, file = TRUE, cleanText = cleanTextDefault)
}
\arguments{
\item{text1}{first text}

\item{text2}{second text}

\item{file}{Whether or not the texts are supplied as character vector
(\code{file=FALSE}) or as file path (the default with \code{file=TRUE}).}

\item{cleanText}{The function to be applied to each line of the text prior
to alignment and difference measurement.}
}
\value{
The function returns a list with items:
\itemize{
  \item \code{text1_orig} and \code{text2_orig} holding the original texts,
  \item \code{text1_clean} and \code{text2_clean} holding the texts after cleaning,
  \item \code{alignment_auto} and \code{alignment_semi} data frames holding
  the corresponding line numbers from \code{text1} and \code{text2}
  and the distance between them
  \item \code{distance_matrix} a matrix of distances between text lines
  -- rows for text1 and columns for text2
 }
}
\description{
A function for aligning text and measuring its distance.
}
\examples{
require(stringr)
text1 = unlist(str_split(
"This part of the
document has stayed the
same from version to
version.  It shouldn't
be shown if it doesn't
change.  Otherwise, that
would not be helping to
compress the size of the
changes.

This paragraph contains
text that is outdated.
It will be deleted in the
near future.

It is important to spell
check this dokument. On
the other hand, a
misspelled word isn't
the end of the world.
Nothing in the rest of
this paragraph needs to
be changed. Things can
be added after it.","\\n"))

text2      = unlist(strsplit(
"This is an important
notice! It should
therefore be located at
the beginning of this
document!

This part of the
document has stayed the
same from version to
version.  It shouldn't
be shown if it doesn't
change.  Otherwise, that
would not be helping to
compress anything.

It is important to spell
check this document. On
the other hand, a
misspelled word isn't
the end of the world.
Nothing in the rest of
this paragraph needs to
be changed. Things can
#be added after it.

This paragraph contains
important new additions
to this document.", "\\n"))

diffr(text1, text2, file=F)
}

