% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffr_to_html.r
\name{diffr_to_html}
\alias{diffr_to_html}
\title{Function for transforming diffr() results to HTML}
\usage{
diffr_to_html(diffrRes, file = F, head = "", foot = "", clean = F)
}
\description{
Function for transforming diffr() results to HTML
}
\keyword{internal}
