% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distanceFunctions.r
\docType{data}
\name{distanceFunctions}
\alias{distanceFunctions}
\title{Set of predefined functions for diffr()}
\format{
An object of class \code{list} of length 4.
}
\usage{
distanceFunctions
}
\description{
Set of functions to calculate distance between each line of first text and
each line of second text.
}
\keyword{datasets}
