% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{example_A1_split}
\alias{example_A1_split}
\title{first text of running example A as character vector}
\format{\preformatted{ chr [1:34] "This part of the" "document has stayed the" ...
}}
\usage{
example_A1_split
}
\description{
first text of running example A as character vector
}
\keyword{datasets}

