% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\docType{data}
\name{example_A2_split}
\alias{example_A2_split}
\title{sample text split into lines}
\format{a text}
\source{
\url{http://en.wikipedia.org\n/w/index.php?title=Diff\n&oldid=622929855}
}
\usage{
example_A2_split
}
\description{
sample text split into lines
}
\keyword{datasets}

