% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatenate_text.r
\name{\%.\%}
\alias{\%.\%}
\title{A function making "pasting" together text a little bit more readable}
\usage{
a \%.\% b
}
\description{
A function making "pasting" together text a little bit more readable
}
\keyword{internal}
