% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlfoot.R
\docType{data}
\name{htmlfoot}
\alias{htmlfoot}
\title{HTML Footer for Rendering Diffs to HTML Page}
\format{
An object of class \code{character} of length 202.
}
\usage{
htmlfoot
}
\description{
HTML Footer for Rendering Diffs to HTML Page
}
\keyword{internal}
