% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlhead.R
\docType{data}
\name{htmlhead}
\alias{htmlhead}
\title{HTML Header for Rendering Diffs to HTML Page}
\format{
An object of class \code{character} of length 18.
}
\usage{
htmlhead
}
\description{
HTML Header for Rendering Diffs to HTML Page
}
\keyword{internal}
