% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{ignoreLinesFunctions}
\alias{ignoreLinesFunctions}
\title{Predefined functions to determine which lines to ignore for distance
calculation and alignment.}
\format{list of functions}
\usage{
ignoreLinesFunctions
}
\arguments{
\item{idep}{ingore empty lines as well as all those starting with '#§#'
(ignore procedure used within IDEP project work)}

\item{empty}{ingore empty lines}
}
\description{
Predefined functions to determine which lines to ignore for distance
calculation and alignment.
}
\keyword{datasets}

