% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ptest}
\alias{ptest}
\title{A simple function that only exists for testing.}
\usage{
ptest()
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
A simple function that only exists for testing.
}
\examples{
ptest()
}

