% Generated by roxygen2 (4.0.2): do not edit by hand
\name{select_function_option}
\alias{select_function_option}
\title{diffr() helper function for options}
\usage{
select_function_option(option, optionList)
}
\arguments{
\item{option}{Parameter to supply the option value to decide upon.}

\item{optionList}{Parameter to supply list of functions to select from (e.g.
\code{cleanTextFunctions}, \code{distanceFunctions},
\code{ignoreLinesFunctions}).}
}
\description{
Function handles character and function inputs. If input is a string the
function looks if it finds corresponding function in the function list (e.g.
in \code{distanceFunctions}, \code{cleanTextFunctions},
\code{ignoreLinesFunctions}). If the input is a function instead it simply
returns taht function.
}
\details{
Returns a function or an error.
}

