% Generated by roxygen2 (4.0.2): do not edit by hand
\name{similarityDefault}
\alias{similarityDefault}
\title{Similarity Default}
\usage{
similarityDefault(textLine1 = "", textLine2 = "")
}
\arguments{
\item{textLine1}{a line of text}

\item{textLine2}{a second line of text}
}
\description{
Similarity Default
}

