% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_alignment_df.r
\name{sort_align_df}
\alias{sort_align_df}
\title{function sorting alignment data frames according to line numbers}
\usage{
sort_align_df(alignDF, text1First = T)
}
\arguments{
\item{alignDF}{the alignment data frame to be sorted}

\item{text1First}{should first text or second text be given priority}
}
\description{
function sorting alignment data frames according to line numbers
}

