% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{text1}
\alias{text1}
\title{this is some extra stuff for developement}
\format{\preformatted{ chr "This part of the\\ndocument has stayed the\\nsame from version to\\nversion.  It shouldn't\\nbe shown if it doesn't\\nchange.  Other"| __truncated__
}}
\usage{
text1
}
\description{
this is some extra stuff for developement
}
\keyword{datasets}

