% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{text2}
\alias{text2}
\title{this is some extra stuff for developement}
\format{\preformatted{ chr "This is an important\\nnotice! It should\\ntherefore be located at\\nthe beginning of this\\ndocument!\\n \\nThis part of the\\ndocume"| __truncated__
}}
\usage{
text2
}
\description{
this is some extra stuff for developement
}
\keyword{datasets}

