% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_align.r
\name{text_align}
\alias{text_align}
\title{Aligning text.}
\usage{
text_align(distM, maxDist = Inf, retMatrix = T)
}
\arguments{
\item{distM}{Distance matrix to base alignment on.}

\item{maxDist}{Maximum distance below which two lines might be considered to
match.}

\item{retMatrix}{Should the result be returned as boolean matrix or as data
frame.}
}
\description{
A helper function for \code{diffr()} that uses a distance matrix to align
text lines.
}
\details{
Gives back an alignment matrix (boolean) or an data frame with line numbers.
}

