% Generated by roxygen2 (4.0.2): do not edit by hand
\name{textAlignAuto}
\alias{textAlignAuto}
\title{Function for Aligning Text Automatically}
\usage{
textAlignAuto(text1, text2, preAlignment, distanceMatrix)

textAlignAuto(text1, text2, preAlignment, distanceMatrix)
}
\arguments{
\item{text1}{}

\item{text2}{}

\item{preAlignment}{}

\item{distanceMatrix}{}

\item{text1}{}

\item{text2}{}

\item{distanceMatrix}{}
}
\description{
Function for Aligning Text Automatically

Function for Aligning Text Semi-Automatically
}

