% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_text_to_alignment.R
\name{add_text_to_alignment}
\alias{add_text_to_alignment}
\title{Helper function for diffr_to_html that combines alignment_df and text}
\usage{
add_text_to_alignment(alignDF, text1, text2)
}
\description{
Helper function for diffr_to_html that combines alignment_df and text
}
\keyword{internal}
