% Generated by roxygen2 (4.0.2): do not edit by hand
\name{align_matrix_to_align_df}
\alias{align_matrix_to_align_df}
\title{Function that transforms an alignment result in matrix form into an alignment result in data frame form}
\usage{
align_matrix_to_align_df(alignM, distM = NULL, text1_indel_dist = NA,
  text2_indel_dist = NA)
}
\arguments{
\item{alignM}{an alingment matrix}

\item{distM}{an distance matrix of same dimensions as alignM}
}
\description{
Function that transforms an alignment result in matrix form into an alignment result in data frame form
}

