% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanTextFunctions.r
\docType{data}
\name{cleanTextFunctions}
\alias{cleanTextFunctions}
\title{Set of predefined functions for diffr()}
\format{list of functions}
\usage{
cleanTextFunctions
}
\description{
Set of functions to clean text before distances are calculated.
}
\keyword{datasets}

