% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{cleanTextFunctions}
\alias{cleanTextFunctions}
\title{Predefined cleanText functions saved in a list.}
\format{list of functions}
\usage{
diffr("ab", "bb", clean="none")
}
\arguments{
\item{none}{does not transform the text at all}

\item{idep}{used in the data gathering of IDEP (Institutional Design in European Parliaments)
working group see citation(diffr) as well as
https://github.com/petermeissner/diffr for further information}
}
\description{
\code{cleanTextFunctions} is a list of predefined functions for text cleansing
to be applied before diffr() calculates distances and alignes text
}
\keyword{datasets}

