% Generated by roxygen2 (4.0.2): do not edit by hand
\name{diffr}
\alias{diffr}
\alias{example_A1,}
\alias{example_A1_split,}
\alias{example_A2,}
\alias{example_A2_split,}
\title{A function for aligning text and measuring its distance.}
\usage{
diffr(text1 = example_A1_split, text2 = example_A2_split, clean = "none",
  dist = "levenshtein", maxDist = "Inf", sortDF = c(1, 2, 0),
  ignore = "empty", ret = "all")
}
\arguments{
\item{text1}{First text. Expected to be a character vector or list with each
  item repesenting a line. Lists will be transformed to character vectors via
  \code{unlist()}. Each vector element of text1 will be compared to each
  element of text2.}

\item{text2}{Second text. See \code{text1}.}

\item{clean}{The function to be applied to each line of the text prior to
  alignment and difference measurement. The parameter might either be a
  string referring to the name of one cleansing function provided by the
  package (see \code{names(cleanTextFunctions)}) or should be a user supplied
  function. Text cleansing functions are expected to accept one character
  vectors as argument and to return a character vector of the very same
  length.}

\item{dist}{The function to be used to calculate distances between text lines
  to be used to do the alignment later on. The parameter might either be a
  string referring to the name of one distance function provided by the
  package (see \code{names(distanceFunctions)}) or should be a user supplied
  function. Distance functions are expected to accept character vectors as
  first and second argument and to return a matrix of distances between text1
  and text2 as intersection between the i_th line and the j_th column -- rows
  represent text1 lines and columns represent text2 lines.}

\item{maxDist}{This parameter holds the maximum distance below that it is
  acceptable for two lines to be considered as matches. maxDist migt either
  be supplied as numeric defining the maximum distance in absolute terms or
  it might be a string combined of a number between 0 and 100 as well as a \%
  character t indicate using a certain percentile of the maximum distance
  found between two lines as cut off point.}

\item{sortDF}{Selects the sorting priority for \code{alignment_df}: Either
  \code{0} for no priority, \code{1} for first text priority, or  \code{2}
  for second text priority.}

\item{ignore}{The function to be used to decide whether or not a particular
  text line should be excluded from calculations -- distances will still be
  calculated (so ignoring lines wont do much about performance) but they are
  neither cleaned nor considered for alignment, furthermore those lines do
  not count for the overall distance of the text. The parameter might either
  be a string referring to the name of one ignore function provided by the
  package (see \code{names(ingoreLinesFunctions)}) or should be a user
  supplied function. Ignore functions are expected to accept one character
  vector as argument and to return a boolean vector of the very same length
  that with TRUE for those lines to be ignored and FALSE otherwise.}

\item{ret}{The parameter defines what should be returned by the function. The
  default value is 'all' while another reasonable value might be 'defaut' to
  get only the alignment data frame. Furthermore, one can supply a vector
  with the names of those objects that should be returned:
  \code{c("text1_orig", "text2_orig", "text1_clean", "text2_clean",
  "distance_matrix", "alignment_df")}.}
}
\value{
The function returns a list with items: \itemize{ \item
  \code{text1_orig} and \code{text2_orig} holding the original texts, \item
  \code{text1_clean} and \code{text2_clean} holding the texts after cleaning,
  \item \code{alignment_df} data frame holding the corresponding line numbers
  from \code{text1} and \code{text2}, the distance between them and the
  corresponding match type\item \code{distance_matrix} a matrix of distances
  between text lines -- rows for text1 and columns for text2 }
}
\description{
This function represents the heart of the diffr package. It can be used to
analyse changes made to texts by comparing two versions of the same text.
}
\details{
The function tries to align unchanged as well as changed lines to each other
and allows to measure the amount of change that occured per line as well as
for the whole text.

Features of the funtion include choosing between text cleansing functions,
distance functions and line ignore functions or to use functions provided by
the user instead. The alignement of the text lines can in addition be tweaked
by the use of the maxDist option that allows for a maximum distance for two
text lines to be considered as matches. Output can be tweaked by three
differing sorting algorithms.
}
\examples{
## diffr(example_A1_split, example_A2_split)
}

