% Generated by roxygen2 (4.0.2): do not edit by hand
\name{diffr}
\alias{diffr}
\title{A function for aligning text and measuring its distance.}
\usage{
diffr(text1, text2)
}
\arguments{
\item{parameter}{1}

\item{parameter}{2}
}
\value{
what is returned again?
}
\description{
A function for aligning text and measuring its distance.
}
\examples{
"example"
}

