% Generated by roxygen2 (4.0.2): do not edit by hand
\name{diffr}
\alias{diffr}
\alias{textAlign}
\title{A function for aligning text and measuring its distance.}
\usage{
diffr(text1, text2, clean = "none", dist = "levenshtein")
}
\arguments{
\item{text1}{first text; expected to be a character vector; each vector
element of text1 will be compared to each element of text2}

\item{text2}{see text1

  (\code{file=FALSE}) or as file path (the default with \code{file=TRUE}).}

\item{clean}{The function to be applied to each line of the text prior to
  alignment and difference measurement.}
}
\value{
The function returns a list with items: \itemize{ \item
  \code{text1_orig} and \code{text2_orig} holding the original texts, \item
  \code{text1_clean} and \code{text2_clean} holding the texts after cleaning,
  \item \code{alignment_auto} and \code{alignment_semi} data frames holding
  the corresponding line numbers from \code{text1} and \code{text2} and the
  distance between them \item \code{distance_matrix} a matrix of distances
  between text lines -- rows for text1 and columns for text2 }
}
\description{
A function for aligning text and measuring its distance.
}
\examples{
require(stringr)
diffr(str_split(exampleA1,"\\n"), str_split(exampleA2,"\\n"))
}

