% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{distanceFunctions}
\alias{distanceFunctions}
\title{Set of predefined functions for diffr()}
\format{\preformatted{List of 4
 $ levenshtein:function (text1, text2)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 10 17 14 3 17 3 10 14
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x00000000094ff120> 
 $ levenwords :function (text1, text2)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 17 17 23 3 17 3 17 23
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x00000000094ff120> 
 $ bow        :function (text1, text2)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 26 11 48 3 11 3 26 48
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x00000000094ff120> 
 $ idep       :function (text1, text2)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 51 11 73 3 11 3 51 73
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x00000000094ff120> 
}}
\usage{
distanceFunctions
}
\description{
Set of functions to calculate distance between each line of first text and
each line of second text.
}
\keyword{datasets}

