% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{distanceFunctions}
\alias{distanceFunctions}
\title{Predefined similarity functions saved in a list.}
\format{list of functions}
\usage{
distanceFunctions
}
\arguments{
\item{nchar}{similarity calculated based on number of characters}

\item{nwords}{similarity calculated based on number of words}

\item{leven}{similarity calculated based on Levenshtein distance (\code{dist()} from \code{proxy} package used for that)}

\item{levenwords}{similarity calculated based on Levenshtein distance that uses words instead of characters as smallest parts (\code{dist()} from \code{proxy} package used for that)}

\item{bow}{similarity calculated based on}

\item{idep}{see bow (distance measure used within IDEP project work)}
}
\description{
\code{similarityFunctions} is a list of predefined functions for text
distance calculation
}
\examples{
names(distanceFunctions)
}
\seealso{
\code{\link[proxy]{pr_DB}} or try: \code{proxy::pr_DB$get_entries(name = NULL, pattern = NULL)}.
}
\keyword{datasets}

