% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_A1.R
\docType{data}
\name{example_A1}
\alias{example_A1}
\title{Example Data Text (A1)}
\format{
An object of class \code{character} of length 1.
}
\usage{
example_A1
}
\description{
Example Data Text (A1)
}
\keyword{internal}
