% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{example_A1}
\alias{example_A1}
\title{first text of running example A}
\format{\preformatted{ chr "This part of the\\ndocument has stayed the\\nsame from version to\\nversion.  It shouldn't\\nbe shown if it doesn't\\nchange.  Other"| __truncated__
}}
\usage{
example_A1
}
\description{
first text of running example A
}
\keyword{datasets}

