% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\docType{data}
\name{example_A1_split}
\alias{example_A1_split}
\title{sample text}
\format{a text split into lines}
\source{
\url{http://en.wikipedia.org\n/w/index.php?title=Diff\n&oldid=622929855}
}
\usage{
example_A1_split
}
\description{
sample text
}
\keyword{datasets}

