% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_A1_split.R
\docType{data}
\name{example_A1_split}
\alias{example_A1_split}
\title{Example Data Text (A1 split)}
\format{
An object of class \code{character} of length 34.
}
\usage{
example_A1_split
}
\description{
Example Data Text (A1 split)
}
\keyword{internal}
