% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{example_A2}
\alias{example_A2}
\title{second text of ruinning example A}
\format{\preformatted{ chr "This is an important\\nnotice! It should\\ntherefore be located at\\nthe beginning of this\\ndocument!\\n \\nThis part of the\\ndocume"| __truncated__
}}
\usage{
example_A2
}
\description{
second text of ruinning example A
}
\keyword{datasets}

