% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_A2.R
\docType{data}
\name{example_A2}
\alias{example_A2}
\title{Example Data Text (A2)}
\format{
An object of class \code{character} of length 1.
}
\usage{
example_A2
}
\description{
Example Data Text (A2)
}
\keyword{internal}
