% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{example_A2_split}
\alias{example_A2_split}
\title{second text of ruinning example A  as character vector}
\format{\preformatted{ chr [1:38] "This is an important" "notice! It should" ...
}}
\usage{
example_A2_split
}
\description{
second text of ruinning example A  as character vector
}
\keyword{datasets}

