% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_A2_split.R
\docType{data}
\name{example_A2_split}
\alias{example_A2_split}
\title{Example Data Text (A2 split)}
\format{
An object of class \code{character} of length 38.
}
\usage{
example_A2_split
}
\description{
Example Data Text (A2 split)
}
\keyword{internal}
