% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ignoreLinesFunctions.r
\docType{data}
\name{ignoreLinesFunctions}
\alias{ignoreLinesFunctions}
\title{Set of predefined functions for diffr()}
\format{An object of class \code{list} of length 3.}
\usage{
ignoreLinesFunctions
}
\description{
Set of functions to determine which lines of text to be ignored.
}
\keyword{datasets}

