% Generated by roxygen2 (4.0.2): do not edit by hand
\name{text_align}
\alias{text_align}
\title{function using tistance matrix to align lines accordingly}
\usage{
text_align(distM, maxDist = Inf, retMatrix = T)
}
\arguments{
\item{distM}{distance matrix to base alignment on}

\item{maxDist}{beyonbd this distance all possible line matches are coded as
deletions and respectively insertions}

\item{retMatrix}{should the result be returned as boolean matrix or as data
frame}
}
\description{
function using tistance matrix to align lines accordingly
}

